+++ src/Libraries/Nop.Core/Domain/Catalog/ProductAttributeValue.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel;
-using System.ComponentModel.DataAnnotations.Schema;
 using Nop.Core.Domain.Localization;
 
 namespace Nop.Core.Domain.Catalog;
@@ -12,17 +11,17 @@ public partial class ProductAttributeValue : BaseEntity, ILocalizedEntity
     /// <summary>
     /// Gets or sets the product attribute mapping identifier
     /// </summary>
-    public int ProductAttributeMappingId { get; set; }
+    public long ProductAttributeMappingId { get; set; }
 
     /// <summary>
     /// Gets or sets the attribute value type identifier
     /// </summary>
-    public int AttributeValueTypeId { get; set; }
+    public long AttributeValueTypeId { get; set; }
 
     /// <summary>
     /// Gets or sets the associated product identifier (used only with AttributeValueType.AssociatedToProduct)
     /// </summary>
-    public int AssociatedProductId { get; set; }
+    public long AssociatedProductId { get; set; }
 
     /// <summary>
     /// Gets or sets the product attribute name
@@ -37,7 +36,7 @@ public partial class ProductAttributeValue : BaseEntity, ILocalizedEntity
     /// <summary>
     /// Gets or sets the picture ID for image square (used with "Image squares" attribute type)
     /// </summary>
-    public int ImageSquaresPictureId { get; set; }
+    public long ImageSquaresPictureId { get; set; }
 
     /// <summary>
     /// Gets or sets the price adjustment (used only with AttributeValueType.Simple)
@@ -95,5 +94,5 @@ public partial class ProductAttributeValue : BaseEntity, ILocalizedEntity
     [EditorBrowsable(EditorBrowsableState.Never)]
     [Browsable(false)]
     [Obsolete("The field is not used since 4.70 and is left only for the update process use the ProductAttributeValuePicture instead")]
-    public int? PictureId { get; set; }
+    public long? PictureId { get; set; }
 }
