+++ src/Libraries/Nop.Core/Domain/Catalog/SpecificationAttributeOption.cs
@@ -10,7 +10,7 @@ public partial class SpecificationAttributeOption : BaseEntity, ILocalizedEntity
     /// <summary>
     /// Gets or sets the specification attribute identifier
     /// </summary>
-    public int SpecificationAttributeId { get; set; }
+    public long SpecificationAttributeId { get; set; }
 
     /// <summary>
     /// Gets or sets the name
