+++ src/Libraries/Nop.Core/Domain/Catalog/TierPrice.cs
@@ -8,17 +8,17 @@ public partial class TierPrice : BaseEntity
     /// <summary>
     /// Gets or sets the product identifier
     /// </summary>
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
 
     /// <summary>
     /// Gets or sets the store identifier (0 - all stores)
     /// </summary>
-    public int StoreId { get; set; }
+    public long StoreId { get; set; }
 
     /// <summary>
     /// Gets or sets the customer role identifier
     /// </summary>
-    public int? CustomerRoleId { get; set; }
+    public long? CustomerRoleId { get; set; }
 
     /// <summary>
     /// Gets or sets the quantity
