+++ src/Libraries/Nop.Core/Domain/Discounts/DiscountManufacturerMapping.cs
@@ -8,5 +8,5 @@ public partial class DiscountManufacturerMapping : DiscountMapping
     /// <summary>
     /// Gets or sets the manufacturer identifier
     /// </summary>
-    public override int EntityId { get; set; }
+    public override long EntityId { get; set; }
 }
