+++ src/Libraries/Nop.Core/Domain/Discounts/DiscountProductMapping.cs
@@ -8,5 +8,5 @@ public partial class DiscountProductMapping : DiscountMapping
     /// <summary>
     /// Gets or sets the product identifier
     /// </summary>
-    public override int EntityId { get; set; }
+    public override long EntityId { get; set; }
 }
