+++ src/Libraries/Nop.Core/Domain/Localization/Language.cs
@@ -40,7 +40,7 @@ public partial class Language : BaseEntity, IStoreMappingSupported
     /// <summary>
     /// Gets or sets the identifier of the default currency for this language; 0 is set when we use the default currency display order
     /// </summary>
-    public int DefaultCurrencyId { get; set; }
+    public long DefaultCurrencyId { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether the language is published
