+++ src/Libraries/Nop.Core/Domain/Security/AclRecord.cs
@@ -8,7 +8,7 @@ public partial class AclRecord : BaseEntity
     /// <summary>
     /// Gets or sets the entity identifier
     /// </summary>
-    public int EntityId { get; set; }
+    public long EntityId { get; set; }
 
     /// <summary>
     /// Gets or sets the entity name
@@ -18,5 +18,5 @@ public partial class AclRecord : BaseEntity
     /// <summary>
     /// Gets or sets the customer role identifier
     /// </summary>
-    public int CustomerRoleId { get; set; }
+    public long CustomerRoleId { get; set; }
 }
