+++ src/Libraries/Nop.Core/Domain/Stores/Store.cs
@@ -56,7 +56,7 @@ public partial class Store : BaseEntity, ILocalizedEntity, ISoftDeletedEntity
     /// <summary>
     /// Gets or sets the identifier of the default language for this store; 0 is set when we use the default language display order
     /// </summary>
-    public int DefaultLanguageId { get; set; }
+    public long DefaultLanguageId { get; set; }
 
     /// <summary>
     /// Gets or sets the display order
